extract attestation-gossip logic to its own module
